# perf: preallocate when listing Gateway API objects in store