@pmalek 2024-04-08
> Out of pure curiosity, how did you spot it?

Just browsing 📖 🤷 

> Would it be worthwhile to submit a patch upstream for the benefit of the entire community?

🤔 Since `ListAll` (and other list functions in [`listers.go`](https://github.com/kubernetes/client-go/blob/v0.29.3/tools/cache/listers.go)) accept an `appendFn`:

```
type AppendFunc func(interface{})
```

It's up to the user to preallocate. We could potentially extend or change the API (but it would be breaking) so that the `ListAll` (or it's sibling) allocates the slices.